refactor: extract pane input helpers
